Changed logic of categorias insertion. Needs verification.

+++ web/remove_cat.php
@@ -16,24 +16,51 @@
                 $db = new PDO("pgsql:host=$host;dbname=$dbname", $user, $password);
                 $db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
 
-                $db->query("begin transaction;");
 
-                $sql = "DELETE FROM super_categoria WHERE nome='$nome_categoria';";
-                echo("<p>Trying to Remove $nome_categoria from super_categoria:</p>");
-                echo("<p>$sql</p>");
-                $db->query($sql);
+                $sql = "SELECT COUNT(categoria) FROM constituida WHERE categoria = '$nome_categoria'";
+                $result = $db->query($sql);
+                $count = $result->fetchColumn();
                 
-                $sql = "DELETE FROM categoria_simples WHERE nome='$nome_categoria';";
-                echo("<p>Trying to Remove $nome_categoria from categoria_simples:</p>");
-                echo("<p>$sql</p>");
-                $db->query($sql);
+                if($count > 0) {
+                    $sql = "SELECT categoria FROM constituida where super_categoria='$supercategoria';";
+                    $result = $db->query($sql);
+                    echo "<p>Não é possível remover $nome_categoria porque está registada como subcategoria de:</p>";
+                    echo "<table>";
+                    foreach($result as $row)
+                    {
+                        echo("<tr>\n");
+                        echo("<td>{$row['categoria']}</td>\n");
+                        echo("</tr>\n");
+                    }
+                    echo "</table>";
 
-                $sql = "DELETE FROM categoria WHERE nome='$nome_categoria';";
-                echo("<p>Trying to Remove $nome_categoria from categoria:</p>");
-                echo("<p>$sql</p>");
-                $db->query($sql);
-
-                $db->query("commit;");
+                    echo "<p>Por favor, remova $nome_categoria como subcategoria desta(s) categorias antes de proceder à sua eliminação.</p>";
+                }
+                else {
+                    $db->query("begin transaction;");
+                    
+                    $sql = "DELETE FROM constituida WHERE super_categoria='$nome_categoria';";
+                    echo("<p>Trying to detach all subcategorias from $nome_categoria:</p>");
+                    echo("<p>$sql</p>");
+                    $db->query($sql);
+    
+                    $sql = "DELETE FROM super_categoria WHERE nome='$nome_categoria';";
+                    echo("<p>Trying to Remove $nome_categoria from super_categoria:</p>");
+                    echo("<p>$sql</p>");
+                    $db->query($sql);
+                    
+                    $sql = "DELETE FROM categoria_simples WHERE nome='$nome_categoria';";
+                    echo("<p>Trying to Remove $nome_categoria from categoria_simples:</p>");
+                    echo("<p>$sql</p>");
+                    $db->query($sql);
+        
+                    $sql = "DELETE FROM categoria WHERE nome='$nome_categoria';";
+                    echo("<p>Trying to Remove $nome_categoria from categoria:</p>");
+                    echo("<p>$sql</p>");
+                    $db->query($sql);
+                
+                    $db->query("commit;");
+                }
 
                 $db = null;
             }
